requirements.txt
streamlit==1.29.0
openai>=1.12.0
pillow==10.1.0
python-dotenv==1.0.0

